feat: replace rollup & babel with tsup
