--- babel.config.js
@@ -1,7 +1,0 @@
-/** @type {import('@babel/core').TransformOptions} */
-export default {
-  presets: [
-    ['@babel/preset-env', { targets: { node: 'current' } }],
-    '@babel/typescript',
-  ],
-};
